+++ Frameworks/UIKit.framework/UICalloutView.h
@@ -70,7 +70,7 @@
 @property(readonly) float UICalloutViewCapPaddingTop;
 @property(readonly) float UICalloutViewHorizontalMargin;
 @property(readonly) float UICalloutViewHorizontalPadding;
-@property(readonly) float UICalloutViewLayoutDuration;
+@property(readonly) double UICalloutViewLayoutDuration;
 @property(readonly) float UICalloutViewMinimumWidth;
 @property(readonly) float UICalloutViewVerticalMargin;
 @property BOOL canAnchorFromBottom;
@@ -95,7 +95,7 @@
 - (float)UICalloutViewCapPaddingTop;
 - (float)UICalloutViewHorizontalMargin;
 - (float)UICalloutViewHorizontalPadding;
-- (float)UICalloutViewLayoutDuration;
+- (double)UICalloutViewLayoutDuration;
 - (float)UICalloutViewMinimumWidth;
 - (float)UICalloutViewVerticalMargin;
 - (void)_fadeViewsIn:(BOOL)arg1;
@@ -116,7 +116,7 @@
 - (void)completeBounceAnimation;
 - (void)dealloc;
 - (id)delegate;
-- (void)fadeOutWithDuration:(float)arg1;
+- (void)fadeOutWithDuration:(double)arg1;
 - (void)getActualAnchorPoint:(struct CGPoint { float x1; float x2; }*)arg1 frame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; }*)arg2 position:(int*)arg3 forDesiredAnchorPoint:(struct CGPoint { float x1; float x2; })arg4 boundaryRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg5;
 - (id)hitTest:(struct CGPoint { float x1; float x2; })arg1 forEvent:(struct __GSEvent { }*)arg2;
 - (id)hitTest:(struct CGPoint { float x1; float x2; })arg1 withEvent:(id)arg2;
